// src/types/project.ts
export interface Project {
  id: string;
  cod: string;
  userId: string;
  createdAt: Date;
  versions: ProjectVersion[];
}

export interface ProjectVersion {
  id: string;
  projectId: string;
  version: string;
  title: string;
  address: string;
  cep: string;
  district: string;
  city: string;
  state: string;
  updatedAt: Date;
  startDate: string;
  constructionStartDate: Date;
  phases: ProjectPhase[];
}

export interface ProjectPhase {
  id: string;
  projectVersionId: string;
  name: string;
  weeks: number;
  isIndependent: boolean;
  startDate: string;
  independentDate: string;
  phaseOrder: number;
  phaseType: PhaseType;
  milestones: Milestone[];
}

export interface Milestone {
  id: string;
  projectPhaseId: string;
  name: string;
  date: Date;
}

export enum PhaseType {
  PROJECT = "PROJECT",
  CONSTRUCTION = "CONSTRUCTION",
}